refactor: Avatar Components
